requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.28.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.3
python-dotenv